+++ web/middleware/books/validate.js
@@ -1,25 +1,32 @@
 const createError = require('http-errors')
 const booksSchema = require('./booksSchema')
 const Joi = require('joi')
-const isDateValid = require('../../lib/validateDate').isDateValid
 const { buildErrorMessages, invalidDateMessage } = require('../../lib/buildMessages')
 
 module.exports = {
   post: async (req, res, next) => {
     try {
       Joi.assert(req.body, booksSchema, {stripUnknown: true, abortEarly: false})
-      isDateValid(req.body.published) ? next() : next(createError(400, invalidDateMessage(req.body.published)))
+      next()
     } catch (err) {
-      next(createError(400, buildErrorMessages(err,'VALIDATION_ERROR')))
+      if (err.details && err.details.some(d => d.type === 'customDate.invalid')) {
+        next(createError(400, invalidDateMessage(req.body.published)))
+      } else {
+        next(createError(400, buildErrorMessages(err, 'VALIDATION_ERROR')))
+      }
     }
   },
 
   put: async (req, res, next) => {
     try {
       Joi.assert(req.body,booksSchema, {stripUnknown: true, abortEarly: false} )
-      isDateValid(req.body.published) ? next() : next(createError(400, invalidDateMessage(req.body.published)))
+      next()
     } catch (err) {
-      next(createError(400, buildErrorMessages(err,'VALIDATION_ERROR')))
+      if (err.details && err.details.some(d => d.type === 'customDate.invalid')) {
+        next(createError(400, invalidDateMessage(req.body.published)))
+      } else {
+        next(createError(400, buildErrorMessages(err, 'VALIDATION_ERROR')))
+      }
     }
   },